Store experiments in the session
