+++ app/views/optimizeScript.scala.html
@@ -6,7 +6,7 @@
     <!--- Optimize scripts, see: https://support.google.com/optimize/answer/7359264 -->
     <!-- Page-hiding snippet -->
     <style>.async-hide {
-        //opacity: 0 !important
+        opacity: 0 !important
     } </style>
     <script async type="text/javascript" src="@assets("optimizeScript.js")"></script>
 }
