+++ assets/components/whySupportVideo/whySupportVideoContainer.js
@@ -6,7 +6,7 @@ import { connect } from 'react-redux';
 
 import WhySupportVideo from 'components/whySupportVideo/whySupportVideo';
 
-import { type CommonState } from 'helpers/page/page';
+import { type CommonState } from 'helpers/page/commonReducer';
 
 
 // ----- Types and Map State/Props ----- //
