+++ assets/helpers/page/page.js
@@ -3,65 +3,28 @@
 // ----- Imports ----- //
 
 import 'ophan';
-import {
-  createStore,
-  combineReducers,
-  applyMiddleware,
-  compose,
-  type Reducer,
-} from 'redux';
-import thunkMiddleware from 'redux-thunk';
 import type { Store } from 'redux';
+import { applyMiddleware, combineReducers, compose, createStore, type Reducer } from 'redux';
+import thunkMiddleware from 'redux-thunk';
 
-import * as abTest from 'helpers/abTests/abtest';
 import type { Participations } from 'helpers/abTests/abtest';
+import * as abTest from 'helpers/abTests/abtest';
 import type { Settings } from 'helpers/settings';
 import * as logger from 'helpers/logger';
 import * as googleTagManager from 'helpers/tracking/googleTagManager';
 import { detect as detectCountry, type IsoCountry } from 'helpers/internationalisation/country';
 import { detect as detectCurrency, type IsoCurrency } from 'helpers/internationalisation/currency';
 import { getAllQueryParamsWithExclusions } from 'helpers/url';
-import {
-  getCampaign,
-  getReferrerAcquisitionData,
-  type Campaign,
-  type ReferrerAcquisitionData,
-} from 'helpers/tracking/acquisitions';
-import {
-  detect as detectCountryGroup,
-  type CountryGroupId,
-} from 'helpers/internationalisation/countryGroup';
-import {
-  type OptimizeExperiments,
-  getOptimizeExperiments,
-  addExperimentUpdateListener, addOptimizeExperiments
-} from 'helpers/tracking/optimize';
+import type { CommonState } from 'helpers/page/commonReducer';
+import { createCommonReducer } from 'helpers/page/commonReducer';
+import { getCampaign, getReferrerAcquisitionData } from 'helpers/tracking/acquisitions';
+import { type CountryGroupId, detect as detectCountryGroup } from 'helpers/internationalisation/countryGroup';
+import { addOptimizeExperiments, readExperimentsFromSession } from 'helpers/optimize/optimize';
 import storeReferrer from 'helpers/tracking/awin';
-
-import { type Action } from './pageActions';
 
 if (process.env.NODE_ENV === 'DEV') {
   import('preact/devtools');
 }
-
-// ----- Types ----- //
-
-export type Internationalisation = {|
-  currencyId: IsoCurrency,
-  countryGroupId: CountryGroupId,
-  countryId: IsoCountry,
-|};
-
-export type CommonState = {
-  campaign: ?Campaign,
-  referrerAcquisitionData: ReferrerAcquisitionData,
-  otherQueryParams: Array<[string, string]>,
-  abParticipations: Participations,
-  settings: Settings,
-  internationalisation: Internationalisation,
-  optimizeExperiments: OptimizeExperiments,
-};
-
 
 // ----- Functions ----- //
 
@@ -91,7 +54,7 @@ function buildInitialState(
   settings: Settings,
 ): CommonState {
   const acquisition = getReferrerAcquisitionData();
-  const optimizeExperiments = getOptimizeExperiments();
+  const optimizeExperiments = readExperimentsFromSession();
   const excludedParameters = ['REFPVID', 'INTCMP', 'acquisitionData'];
   const otherQueryParams = getAllQueryParamsWithExclusions(excludedParameters);
   const internationalisation = {
@@ -108,30 +71,6 @@ function buildInitialState(
     abParticipations,
     settings,
     optimizeExperiments,
-  };
-
-}
-
-// Sets up the common reducer with its initial state.
-function createCommonReducer(initialState: CommonState): (state?: CommonState, action: Action) => CommonState {
-
-  return function commonReducer(
-    state?: CommonState = initialState,
-    action: Action,
-  ): CommonState {
-
-    switch (action.type) {
-
-      case 'SET_COUNTRY':
-        return {
-          ...state,
-          internationalisation: { ...state.internationalisation, countryId: action.country },
-        };
-
-      default:
-        return state;
-    }
-
   };
 
 }
@@ -196,7 +135,6 @@ function init<S, A>(
 // ----- Exports ----- //
 
 export {
-  createCommonReducer,
   init,
   statelessInit,
   doNotTrack,
