+++ assets/helpers/utilities.js
@@ -77,6 +77,7 @@ function deserialiseJsonObject(serialised: string): ?Object {
 
 }
 
+
 // Adds leading zeros
 function addLeadingZeros(value: number, length: number = 2): string {
   let valueStr = String(value);
