+++ assets/pages/digital-subscription-checkout/digitalSubscriptionCheckoutReducer.js
@@ -2,7 +2,7 @@
 
 // ----- Imports ----- //
 
-import type { CommonState } from 'helpers/page/page';
+import type { CommonState } from 'helpers/page/commonReducer';
 import { type IsoCountry, fromString } from 'helpers/internationalisation/country';
