+++ assets/pages/support-landing/components/subscriptionsSection.jsx
@@ -9,7 +9,7 @@ import { getSubsLinks } from 'helpers/externalLinks';
 import { getCampaign, type ReferrerAcquisitionData } from 'helpers/tracking/acquisitions';
 import { type CountryGroupId } from 'helpers/internationalisation/countryGroup';
 import { type Participations } from 'helpers/abTests/abtest';
-import { type OptimizeExperiments } from 'helpers/tracking/optimize';
+import { type OptimizeExperiments } from 'helpers/optimize/optimize';
 
 import ThreeSubscriptions from 'components/threeSubscriptions/threeSubscriptions';
 import DigitalPack from 'components/subscriptionBundles/digitalPack';
